refactor(test-runner): extract shared client/session setup
Pull the per-invocation session client and ClickHouse-settings assembly
out of the passthrough backend into backends/shared.ts so a second
backend can reuse them. No behaviour change.
