+++ tests/clickhouse-test-runner/src/backends/shared.ts
@@ -1,0 +1,106 @@
+import { randomUUID } from "node:crypto";
+import {
+  ClickHouseLogLevel,
+  createClient,
+  type ClickHouseClient,
+} from "@clickhouse/client";
+import type { ParsedArgs } from "../args.js";
+import { appendLog } from "../log.js";
+import { errorMatchesExpectation, type Statement } from "../test-hint.js";
+
+export interface BackendOptions {
+  args: ParsedArgs;
+  statements: Statement[];
+  logPath: string;
+}
+
+/** Server settings to forward, assembled the same way for every backend. */
+export function buildClickHouseSettings(
+  args: ParsedArgs,
+): Record<string, string | number> {
+  const settings: Record<string, string | number> = {};
+  settings["default_format"] = "TabSeparated";
+  if (args.logComment !== null && args.logComment.length > 0) {
+    settings["log_comment"] = args.logComment;
+  }
+  if (args.sendLogsLevel !== null && args.sendLogsLevel.length > 0) {
+    settings["send_logs_level"] = args.sendLogsLevel;
+  }
+  if (args.maxInsertThreads !== null && args.maxInsertThreads.length > 0) {
+    settings["max_insert_threads"] = args.maxInsertThreads;
+  }
+  for (const [k, v] of Object.entries(args.serverSettings)) {
+    settings[k] = v;
+  }
+  return settings;
+}
+
+/**
+ * Create a client bound to a dedicated per-invocation `session_id` so that
+ * settings applied via `SET ...` in one statement persist for subsequent
+ * statements within the same .sql script. Without a session, every
+ * `client.exec(...)` is an independent HTTP request and `SET` has no effect on
+ * later requests, which breaks upstream tests that rely on patterns like
+ *   SET allow_deprecated_syntax_for_merge_tree = 1;
+ *   CREATE TABLE ... ENGINE = MergeTree(d, k, 8192);
+ */
+export function createSessionClient(
+  args: ParsedArgs,
+  logPath: string,
+): ClickHouseClient {
+  const proto = args.secure ? "https" : "http";
+  const url = `${proto}://${args.host}:${args.port}`;
+  const sessionId = `clickhouse-js-test-runner-${randomUUID()}`;
+  appendLog(logPath, "session_id=" + sessionId);
+  return createClient({
+    url,
+    username: args.user,
+    password: args.password,
+    database: args.database,
+    session_id: sessionId,
+    // The client logs request errors to stderr by default. We surface failures
+    // ourselves (and deliberately swallow errors matched by a `-- { serverError
+    // ... }` hint), and upstream `clickhouse-test` fails any test that writes to
+    // stderr — so keep the client itself silent.
+    log: { level: ClickHouseLogLevel.OFF },
+  });
+}
+
+/**
+ * Reconcile the outcome of a single statement against its expected-error hint,
+ * the same way for every backend:
+ *   - hinted (`-- { serverError ... }`): an error is the success path, its
+ *     absence (or a mismatching error) is a failure;
+ *   - un-hinted: any error propagates.
+ * Throws on failure; returns normally when the statement is considered passed.
+ */
+export function settleExpectedError(
+  stmt: Statement,
+  execError: unknown,
+  logPath: string,
+): void {
+  const expected = stmt.expectedError;
+  if (expected !== null) {
+    if (execError === null) {
+      throw new Error(
+        `Expected error (${expected.label}) but the query succeeded: ${stmt.sql}`,
+      );
+    }
+    if (!errorMatchesExpectation(execError, expected)) {
+      const actual =
+        execError instanceof Error ? execError.message : String(execError);
+      throw new Error(
+        `Expected error (${expected.label}) but got a different error: ${actual}`,
+      );
+    }
+    appendLog(logPath, "expected_error_matched=" + expected.label);
+    return;
+  }
+
+  if (execError !== null) {
+    const msg =
+      execError instanceof Error ? execError.message : String(execError);
+    appendLog(logPath, "error=" + msg);
+    throw execError;
+  }
+}
